added column ordering, and readded visibility toggle
will clean up code later?
