--- src/components/ColumnVisibilityState.tsx
@@ -1,9 +1,0 @@
-export interface ColumnVisibilityState {
-  id: boolean;
-  name: boolean;
-  age: boolean;
-  id2: boolean;
-  name2: boolean;
-  age2: boolean;
-  [key: string]: boolean;
-}
