--- src/components/Table2.tsx
@@ -1,114 +1,0 @@
-import {
-  flexRender,
-  getCoreRowModel,
-  useReactTable,
-} from "@tanstack/react-table";
-import { useState } from "react";
-import dataJSON from "./mockdata.json";
-import { createColumnHelper } from "@tanstack/react-table";
-import Checkbox from "./Checkbox";
-
-export type Row = {
-  id: string;
-  first_name: string;
-  last_name: string;
-};
-
-export type RowSelectionState = Record<string, boolean>;
-
-export type RowSelectionTableState = {
-  rowSelection: RowSelectionState;
-};
-
-function Table() {
-  const [rows, setRows] = useState(dataJSON);
-  const [rowSelection, setRowSelection] = useState<RowSelectionState>({});
-  const columnHelper = createColumnHelper<Row>();
-  const colDef = [
-    {
-      id: "select",
-      header: ({ table }) => (
-        <Checkbox
-          {...{
-            checked: table.getIsAllRowsSelected(),
-            indeterminate: table.getIsSomeRowsSelected(),
-            onChange: table.getToggleAllRowsSelectedHandler(),
-          }}
-        />
-      ),
-      cell: ({ row }) => (
-        <Checkbox
-          {...{
-            checked: row.getIsSelected(),
-            disabled: !row.getCanSelect(),
-            indeterminate: row.getIsSomeSelected(),
-            onChange: row.getToggleSelectedHandler(),
-          }}
-        />
-      ),
-    },
-    columnHelper.accessor("id", {
-      header: () => "ID",
-      cell: (info) => info.getValue(),
-    }),
-    columnHelper.accessor("first_name", {
-      header: () => "First Name",
-      cell: (info) => info.getValue(),
-    }),
-    columnHelper.accessor("last_name", {
-      header: () => "Last Name",
-      cell: (info) => info.getValue(),
-    }),
-  ];
-  const table = useReactTable({
-    data: rows,
-    columns: colDef,
-    getCoreRowModel: getCoreRowModel(),
-    state: {
-      rowSelection,
-    },
-    onRowSelectionChange: setRowSelection,
-    enableRowSelection: true,
-  });
-  /*const columns = [
-    
-  ]; */
-  return (
-    <div className="relative overflow-x-auto shadow-md sm:rounded-lg">
-      <table className="w-full text-left text-sm text-gray-500 dark:text-gray-400 rtl:text-right">
-        <thead className="bg-gray-50 text-xs uppercase text-gray-700 dark:bg-gray-700 dark:text-gray-400">
-          {table.getHeaderGroups().map((headerGroup) => (
-            <tr key={headerGroup.id}>
-              {headerGroup.headers.map((header) => (
-                <th key={header.id} className="px-6 py-3">
-                  <div class="flex items-center">
-                    {flexRender(
-                      header.column.columnDef.header,
-                      header.getContext(),
-                    )}
-                  </div>
-                </th>
-              ))}
-            </tr>
-          ))}
-        </thead>
-        <tbody>
-          {table.getRowModel().rows.map((row) => (
-            <tr
-              key={row.id}
-              className="border-b bg-white hover:bg-gray-50 dark:border-gray-700 dark:bg-gray-800 dark:hover:bg-gray-600"
-              onClick={row.getToggleSelectedHandler()}
-            >
-              {row.getVisibleCells().map((cell) => (
-                <td key={cell.id} className="px-6 py-4">
-                  {flexRender(cell.column.columnDef.cell, cell.getContext())}
-                </td>
-              ))}
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    </div>
-  );
-}
-export default Table;
