+++ src/main.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import "./globals.css";
-import Page from "./components/Page";
+import Table from "./components/Table";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
-    <Page />
+    <Table />
   </React.StrictMode>,
 );
